Sort and align a ledger-cli journal file by date, payee, and account.
Line item accounts are sorted by account name, but I think the item
with the largest absolute-value amount should go first.  This item's
value is used when the transaction is folded down in vim.  It's
misleading when it's not the full amount of the transaction.

Account names on line items are left aligned, taking the status (or
absence) into consideration.  Amounts are right aligned.  This makes
the decimal points align if they contain the same precision.  A future
refinement may be to line up the decimal points regardless of precision.